+++ cdk/constructs/ssm_document_lustre_preload.py
@@ -2,9 +2,10 @@
 # SPDX-License-Identifier: Apache-2.0,
 import logging
 import os
-
+import hashlib
 import yaml
 from aws_cdk import Aws as aws
+import aws_cdk as cdk
 from aws_cdk import Fn as fn
 from aws_cdk import aws_ec2 as ec2
 from aws_cdk import aws_fsx as fsx
@@ -38,13 +39,33 @@ class SSMDocumentLustrePreload(Construct):
         ec2_ami = ec2.MachineImage.from_ssm_parameter(
             "/aws/service/ecs/optimized-ami/amazon-linux-2/recommended/image_id"
         )
-
+        # EC2 Role
+        ec2_ssm_role = iam.Role(
+            self,
+            "ssm-ec2-instance-role",
+            description="AWS Batch with FFMPEG : Enables EC2 to access SSM",
+            assumed_by=iam.CompositePrincipal(
+                iam.ServicePrincipal("ec2.amazonaws.com"),
+            ),
+            managed_policies=[
+                iam.ManagedPolicy.from_aws_managed_policy_name(
+                    "service-role/AmazonEC2RoleforSSM"
+                ),
+                iam.ManagedPolicy.from_aws_managed_policy_name(
+                    "AmazonSSMManagedInstanceCore"
+                ),
+            ],
+        )
+        ec2_ssm_instance_profile = iam.InstanceProfile(
+            self,
+            "ssm-ec2-instance-profile",
+            role=ec2_ssm_role,
+        )
         # IAM assume role for the execution of the SSM Automation
         ssm_assume_role = iam.Role(
             self,
             "ssm-assume-role",
             description="AWS Batch with FFMPEG : SSM assume role",
-            role_name="batch-ffmpeg-ssm-assume-role",
             assumed_by=iam.CompositePrincipal(
                 iam.ServicePrincipal("ssm.amazonaws.com"),
             ),
@@ -65,37 +86,12 @@ class SSMDocumentLustrePreload(Construct):
                                 + aws.PARTITION
                                 + ":iam::"
                                 + aws.ACCOUNT_ID
-                                + ":role/batch-ffmpeg-ssm-ec2-instance-role"
+                                + f":role/{ec2_ssm_role.role_name}"
                             ],
                         )
                     ]
                 )
             },
-        )
-
-        # EC2 Role
-        ec2_ssm_role = iam.Role(
-            self,
-            "ssm-ec2-instance-role",
-            description="AWS Batch with FFMPEG : Enables EC2 to access SSM",
-            role_name="batch-ffmpeg-ssm-ec2-instance-role",
-            assumed_by=iam.CompositePrincipal(
-                iam.ServicePrincipal("ec2.amazonaws.com"),
-            ),
-            managed_policies=[
-                iam.ManagedPolicy.from_aws_managed_policy_name(
-                    "service-role/AmazonEC2RoleforSSM"
-                ),
-                iam.ManagedPolicy.from_aws_managed_policy_name(
-                    "AmazonSSMManagedInstanceCore"
-                ),
-            ],
-        )
-        ec2_ssm_instance_profile = iam.InstanceProfile(
-            self,
-            "ssm-ec2-instance-profile",
-            instance_profile_name="batch-ffmpeg-ssm-ec2-instance-profile",
-            role=ec2_ssm_role,
         )
 
         # EC2 UserData
@@ -112,6 +108,8 @@ class SSMDocumentLustrePreload(Construct):
             )
 
         user_data_lustre_base64 = fn.base64(user_data_lustre_txt)
+        md5_hash = hashlib.md5(user_data_lustre_txt.encode(), usedforsecurity=False)
+        short_hash = md5_hash.hexdigest()[:5]
 
         # SSM Automation Document YAML
         with open(from_root("cdk", "ssm-documents", "lustre-preload.yaml")) as f:
@@ -131,6 +129,9 @@ class SSMDocumentLustrePreload(Construct):
             lustre_preload_document_yaml = lustre_preload_document_yaml.replace(
                 "%INSTANCE_PROFILE_ARN%", ec2_ssm_instance_profile.instance_profile_arn
             )
+            lustre_preload_document_yaml = lustre_preload_document_yaml.replace(
+                "%HASH_USERDATA%", short_hash
+            )
 
             self.ssm_document = ssm.CfnDocument(
                 self,
@@ -140,4 +141,5 @@ class SSMDocumentLustrePreload(Construct):
                 document_format="YAML",
                 name="batch-ffmpeg-lustre-preload",
                 update_method="NewVersion",
+                tags=[cdk.CfnTag(key="hash-user-data", value=short_hash)],
             )